# Inline CacheTableInternal methods